fixed-- - sign up make profile picture not a must--fixed - when giving or removing role check if the user already have the role---just work with current impl - delete user has an error-- issue on role - profile image just update user profile image field--- done

+++ src/main/java/com/brandyodhiambo/bibleApi/feature/usermgt/controller/UserController.java
@@ -1,17 +1,16 @@
 package com.brandyodhiambo.bibleApi.feature.usermgt.controller;
 
-import com.brandyodhiambo.bibleApi.feature.usermgt.models.UserPrincipal;
 import com.brandyodhiambo.bibleApi.feature.usermgt.models.dto.UserDto;
 import com.brandyodhiambo.bibleApi.feature.usermgt.service.user.UserService;
 import com.brandyodhiambo.bibleApi.feature.usermgt.models.Users;
 import com.brandyodhiambo.bibleApi.util.ApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/users")
@@ -41,18 +40,26 @@ public class UserController {
     public ResponseEntity<Users> updateUser(
             @PathVariable String username,
             @RequestBody Users updatedUser,
-            @AuthenticationPrincipal UserPrincipal currentUser) {
+            @AuthenticationPrincipal Users currentUser) {
         Users user = userService.updateUser(updatedUser, username, currentUser);
         return ResponseEntity.ok(user);
     }
 
     @DeleteMapping("/delete/user/{username}")
-    public ResponseEntity<ApiResponse> deleteUser(
-            @PathVariable String username,
-            @AuthenticationPrincipal UserPrincipal currentUser) {
-        ApiResponse response = userService.deleteUser(username, currentUser);
-        return ResponseEntity.ok(response);
+    public ResponseEntity<?> deleteUser(@PathVariable String username, @AuthenticationPrincipal Users currentUser) {
+        if (currentUser == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ApiResponse(false, "You are not authorized to delete this user"));
+        }
+
+        if (!currentUser.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ApiResponse(false, "You do not have permission to delete this user" +currentUser.getAuthorities()));
+        }
+
+        // Proceed with user deletion
+        userService.deleteUser(username, currentUser);
+        return ResponseEntity.ok(new ApiResponse(true, "User deleted successfully"));
     }
+
 
     @PostMapping("/{username}/giveAdmin")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
